requirements.txt
# Python dependencies — pin versions
# Core
#python>=3.8,<3.11

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Anomaly Detection
anomalib>=1.0.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
albumentations>=1.3.0
Pillow>=10.0.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Experiment Tracking
wandb>=0.15.0
tensorboard>=2.14.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Progress Bar
tqdm>=4.66.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Utilities
pathlib>=1.0.1
typing-extensions>=4.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Streamlit (for demo)
streamlit>=1.28.0

# Optional: WinCLIP dependencies
# clip @ git+https://github.com/openai/CLIP.git
# ftfy>=6.1.0
# regex>=2023.8.8

# Optional: For medical datasets
# nibabel>=5.1.0
# pydicom>=2.4.0


# Week 4-5: Additional dependencies
timm>=0.9.0  # For model backbones
pytorch-lightning>=2.0.0  # For Anomalib
torchmetrics>=1.0.0  # For metrics
